backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9

# Authentication & Security
pyjwt==2.8.0
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
passlib==1.7.4

# Redis & Caching
redis==5.0.1
hiredis==2.3.2

# AI Integration
google-generativeai==0.8.0

# Async & Performance
asyncpg==0.29.0
aioredis==2.0.1
httpx==0.26.0

# Email
aiosmtplib==3.0.1
email-validator==2.1.0

# Utils
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Docker SDK (for lab management)
docker==7.0.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# Development
black==23.12.1
flake8==7.0.0
mypy==1.8.0
